IO port driver minor fixes

+++ AltairHL_emulator/PortDrivers/onboard_io.c
@@ -94,7 +94,7 @@ DX_ASYNC_HANDLER(async_accelerometer_stop_handler, handle)
 }
 DX_ASYNC_HANDLER_END
 
-size_t onboard_output(int port, int data, char *buffer, size_t buffer_length)
+size_t onboard_output(int port, uint8_t data, char *buffer, size_t buffer_length)
 {
     size_t len = 0;
